requirements.txt
Package            Version
------------------ ------------------------
certifi            2022.6.15
charset-normalizer 2.1.0
click              7.1.2
cycler             0.11.0
fonttools          4.34.4
idna               3.3
imageio            2.19.5
Jinja2             2.11.2
joblib             1.1.0
kiwisolver         1.4.4
kornia             0.6.0
MarkupSafe         2.0.1
matplotlib         3.5.2
msgpack-python     0.5.6
networkx           2.8.5
numpy              1.23.1
opencv-python      4.6.0.66
packaging          21.3
pandas             1.4.3
Pillow             9.2.0
pip                22.1.2
PyOpenGL           3.1.5
pyparsing          3.0.9
python-dateutil    2.8.2
pytz               2022.1
PyWavelets         1.3.0
requests           2.28.1
scikit-image       0.19.3
scikit-learn       1.1.1
scipy              1.4.1
setuptools         57.0.0
Shapely            1.8.2
six                1.16.0
threadpoolctl      3.1.0
tifffile           2022.5.4
torch              1.13.0.dev20220720+cu116
torchvision        0.14.0.dev20220720+cu116
typing_extensions  4.3.0
urllib3            1.26.10
wheel              0.36.2
